Implemented different mass for each item and Damage Types.

+++ src/main/java/tgw/evolution/items/IMass.java
@@ -2,5 +2,5 @@ package tgw.evolution.items;
 
 public interface IMass {
 
-    int getMass();
+    double getMass();
 }
